[정세한 - teamA] Assignment : 인증 & 인가 적용+ ErrorHandling

+++ teamA/saehan/services/postService.js
@@ -1,4 +1,6 @@
 const postDao = require("../models/postDao");
+const userDao =require('../models/userDao');
+const jwt = require('jsonwebtoken');
 
 const allPosts = async () => {
     const postsList = await postDao.postsList();
@@ -6,17 +8,37 @@ const allPosts = async () => {
     return postsList;
 };
 
-const posting = async(title, content, user_id) => {
+const posting = async(title, content, user_id, token) => {
+    if(typeof title != 'string' || typeof content != 'string' || typeof parseInt(user_id) != 'number'){
+        const err = new Error('INVALID_INPUT_TYPE');
+        err.statusCode = 409;
+        throw err;
+    };
+
+    const decode =jwt.verify(token, process.env.TOKEN_SECRET);
+    const userId = Object.values(decode)[0];
+    
+    if(user_id != userId){
+        const err = new Error('INVALID_INPUT');
+        err.statusCode = 400;
+        throw err;
+    };
+
     const postUp = await postDao.posting(
         title,
         content,
         user_id
-    );
-
+        );     
     return postUp;
 };
 
 const update = async(userId,postId,title,content) => {
+    if(typeof parseInt(userId) != 'number' || typeof parseInt(postId) != 'number' || typeof title != 'string' || typeof content != 'string'){
+        const err = new Error('INVALID_INPUT_TYPE');
+        err.statusCode = 409;
+        throw err;
+    };
+    
     const updatePost = await postDao.updatePost(
         userId,
         postId,
@@ -28,6 +50,12 @@ const update = async(userId,postId,title,content) => {
 };
 
 const deletePost = async(postId) => {
+    if(typeof(postId) != number){
+        const err = new Error('INVALID_INPUT_TYPE');
+        err.statusCode = 409;
+        throw err;
+    };
+    
     const removePost = await postDao.removePost(
         postId
     );
@@ -40,6 +68,4 @@ module.exports = {
     posting,
     update,
     deletePost
-
-
 }
